+++ rmk-config/src/lib.rs
@@ -560,7 +560,7 @@ pub struct MorsesConfig {
 #[derive(Clone, Debug, Deserialize)]
 #[serde(deny_unknown_fields)]
 pub struct MorseConfig {
-    // name of tap hold profile (to address BehaviorConfig::morse.profiles[self.profile])
+    // name of morse profile (to address BehaviorConfig::morse.profiles[self.profile])
     pub profile: Option<String>,
 
     pub tap: Option<String>,