+++ rmk/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactor tap-hold, and introduced morse_actions to tap-dance to support real morse code like tap/hold patterns 
-- Positional key and tap hold profile configuration introduced
+- Positional and per key morse profile configuration introduced for tap hold like, morse like keys
 - Rename chordal tap to unilateral tap
 - Rewrite led indicator, use controller system